tutina-app/pyproject.toml
[tool.poetry]
name = "tutina-app"
version = "0.1.0"
description = "Web application serving predictions using the Tutina AI model"
authors = ["Jaakko Moisio <jaakko@moisio.fi>"]
readme = "README.md"
packages = [{ include = "tutina" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
pydantic = "^2.10.3"
tutina-lib = { path="../tutina-lib", develop=true }
tutina-ai = { path="../tutina-ai", develop=true }
dict-deep = "^4.1.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.3"
mypy = "^1.13.0"
pandas-stubs = "^2.2.3.241126"
pytest = "^8.3.4"
aiosqlite = "^0.20.0"
pytest-asyncio = "^0.25.0"
faker = "^33.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
select = ["I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
